Fix: fixed the import issue by agent chat of 10k product
